+++ src/scene/triangle.cpp
@@ -35,7 +35,7 @@ bool Triangle::has_intersection(const Ray &r) const {
 		-r.d.y, e1.y, e2.y,
 		-r.d.z, e1.z, e2.z);
 	Vector3D tuv(mat.inv() * (r.o - p1));
-	float t(tuv.x), u(tuv.y), v(tuv.z), w(1.0 - u - v);
+	double t(tuv.x), u(tuv.y), v(tuv.z), w(1.0 - u - v);
 	if (!((u >= 0.0) && (v >= 0.0) && (w >= 0.0))) return false;
 	if (!((t > r.min_t) && (t < r.max_t))) return false;
 	r.max_t = t;
@@ -56,7 +56,7 @@ bool Triangle::intersect(const Ray &r, Intersection *isect) const {
 		-r.d.y, e1.y, e2.y,
 		-r.d.z, e1.z, e2.z);
 	Vector3D tuv(mat.inv() * (r.o - p1));
-	float t(tuv.x), u(tuv.y), v(tuv.z), w(1.0 - u - v);
+	double t(tuv.x), u(tuv.y), v(tuv.z), w(1.0 - u - v);
 	if (!((u >= 0.0) && (v >= 0.0) && (w >= 0.0))) return false;
 	if (!((t > r.min_t) && (t < r.max_t))) return false;
 	r.max_t = t;
